pyproject.toml
[tool.poetry]
name = "hierarchical"
version = "0.1.0"
description = ""
authors = ["Cory Schlaudecker <cory.schlaudecker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
